# lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * YTDown API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryItem {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  format: string;
  /** @nullable */
  filesize?: number | null;
  status: string;
  createdAt: string;
}