# packages/api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * YTDown API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VideoUrlInput {
  /** YouTube video URL */
  url: string;
}

export interface VideoFormat {
  formatId: string;
  ext: string;
  quality: string;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  filesize?: number | null;
  /** @nullable */
  vcodec?: string | null;
  /** @nullable */
  acodec?: string | null;
  /** @nullable */
  fps?: number | null;
  /** @nullable */
  tbr?: number | null;
  /** Logical container group for UI grouping (mp4, webm, audio) */
  container?: string;
  hasAudio?: boolean;
  /**
     * Target bitrate in kbps for synthetic MP3 conversion options
     * @nullable
     */
  audioBitrate?: number | null;
}

export interface VideoInfo {
  id: string;
  title: string;
  url: string;
  formats: VideoFormat[];
  /** @nullable */
  duration: number | null;
  thumbnail: string;
  uploader: string;
  /** @nullable */
  viewCount: number | null;
  /** @nullable */
  uploadDate: string | null;
  /** @nullable */
  description?: string | null;
  ffmpegAvailable?: boolean;
}

export interface DownloadRequest {
  url: string;
  formatId: string;
  audioOnly?: boolean;
  /**
     * Target MP3 bitrate in kbps (320, 256, 192, or 128)
     * @nullable
     */
  audioBitrate?: number | null;
}

export type DownloadJobStatus = typeof DownloadJobStatus[keyof typeof DownloadJobStatus];


export const DownloadJobStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  failed: 'failed',
} as const;

export interface DownloadJob {
  jobId: string;
  status: DownloadJobStatus;
  url: string;
  formatId: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  filename?: string | null;
  /** @nullable */
  error?: string | null;
  createdAt: string;
}

export interface HistoryItem {
  id: number;
  url: string;
  title: string;
  /** @nullable */
  thumbnail?: string | null;
  format: string;
  /** @nullable */
  filesize?: number | null;
  status: string;
  createdAt: string;
}

export interface HistoryList {
  items: HistoryItem[];
  total: number;
}

export type HistoryStatsPopularFormatsItem = {
  format: string;
  count: number;
};

export interface HistoryStats {
  totalDownloads: number;
  successfulDownloads: number;
  totalSizeBytes: number;
  popularFormats: HistoryStatsPopularFormatsItem[];
}

export type SupportedFormatType = typeof SupportedFormatType[keyof typeof SupportedFormatType];


export const SupportedFormatType = {
  video: 'video',
  audio: 'audio',
} as const;

export interface SupportedFormat {
  id: string;
  label: string;
  ext: string;
  type: SupportedFormatType;
  description?: string;
}

export type GetHistoryParams = {
limit?: number;
offset?: number;
};