# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YTDown API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history:
    get:
      operationId: getHistory
      tags: [history]
      summary: Get download history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Download history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryList"
  /history/{id}:
    delete:
      operationId: deleteHistoryItem
      tags: [history]
      summary: Delete a history item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    HistoryItem:
      type: object
      required:
        - id
        - url
        - title
        - format
        - status
        - createdAt
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
        format:
          type: string
        filesize:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
    HistoryList:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HistoryItem"
        total:
          type: integer
...