# packages/api-zod/src/generated/types/historyList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * YTDown API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryItem } from './historyItem';

export interface HistoryList {
  items: HistoryItem[];
  total: number;
}